rename: utils 패키지 이름을 단수형(util)으로 변경
