+++ packy-api/src/main/java/com/dilly/gift/application/GiftService.java
@@ -4,15 +4,15 @@ import com.dilly.gift.adaptor.GiftBoxReader;
 import com.dilly.gift.adaptor.LetterReader;
 import com.dilly.gift.adaptor.PhotoReader;
 import com.dilly.gift.adaptor.ReceiverReader;
+import com.dilly.gift.domain.Photo;
 import com.dilly.gift.domain.giftbox.GiftBox;
 import com.dilly.gift.domain.letter.Letter;
-import com.dilly.gift.domain.Photo;
 import com.dilly.gift.domain.receiver.Receiver;
 import com.dilly.gift.dto.response.GiftResponseDto.ItemResponse;
 import com.dilly.gift.dto.response.LetterResponse;
 import com.dilly.gift.dto.response.MusicResponse;
 import com.dilly.gift.dto.response.PhotoResponseDto.PhotoWithoutSequenceResponse;
-import com.dilly.global.utils.SecurityUtil;
+import com.dilly.global.util.SecurityUtil;
 import com.dilly.member.adaptor.MemberReader;
 import com.dilly.member.domain.Member;
 import jakarta.transaction.Transactional;
